Aniruddha_Bose/Group_12/Hackathon/requirements.txt
langgraph>=0.2
langchain-openai
langchain-core
pydantic>=2.0
python-dotenv
requests
tavily-python
rich



langchain
langchain-community

lancedb

pypdf
python-docx
tiktoken
openai
